+++ app/src/main/java/com/itec/donelio/di/RepositoryModule.kt
@@ -1,12 +1,16 @@
 package com.itec.donelio.di
 
+import com.itec.donelio.data.repository.CampaniaInsumoRepositoryImpl
 import com.itec.donelio.data.repository.CampaniaRepositoryImpl
 import com.itec.donelio.data.repository.CosechaRepositoryImpl
 import com.itec.donelio.data.repository.InsumoRepositoryImpl
+import com.itec.donelio.data.repository.ObservacionRepositoryImpl
 import com.itec.donelio.data.repository.TareaRepositoryImpl
+import com.itec.donelio.domain.repository.CampaniaInsumoRepository
 import com.itec.donelio.domain.repository.CampaniaRepository
 import com.itec.donelio.domain.repository.CosechaRepository
 import com.itec.donelio.domain.repository.InsumoRepository
+import com.itec.donelio.domain.repository.ObservacionRepository
 import com.itec.donelio.domain.repository.TareaRepository
 import dagger.Binds
 import dagger.Module
@@ -41,4 +45,16 @@ abstract class RepositoryModule {
     abstract fun bindInsumoRepository(
         impl: InsumoRepositoryImpl
     ): InsumoRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindCampaniaInsumoRepository(
+        impl: CampaniaInsumoRepositoryImpl
+    ): CampaniaInsumoRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindObservacionRepository(
+        impl: ObservacionRepositoryImpl
+    ): ObservacionRepository
 }