+++ app/src/main/java/com/itec/donelio/domain/model/Campania.kt
@@ -4,5 +4,6 @@ data class Campania(
     val id: Int,
     val nombre: String,
     val fechaInicio: Long,
-    val estaActiva: Boolean
+    val estaActiva: Boolean,
+    val cultivo: String
 )