+++ app/src/main/java/com/itec/donelio/domain/repository/CosechaRepository.kt
@@ -4,7 +4,7 @@ import com.itec.donelio.domain.model.Cosecha
 import kotlinx.coroutines.flow.Flow
 
 interface CosechaRepository {
-    fun getCosechasByCampania(campaniaId: Int): Flow<List<Cosecha>>
+    fun getCosechasByCampania(idCampania: Int): Flow<List<Cosecha>>
 
     suspend fun getCosechaById(id: Int): Cosecha?