+++ app/src/main/java/com/itec/donelio/domain/use_case/EliminarCampaniaUseCase.kt
@@ -1,7 +1,12 @@
 package com.itec.donelio.domain.use_case
 
 import com.itec.donelio.domain.model.Campania
+import com.itec.donelio.domain.model.Resource
 import com.itec.donelio.domain.repository.CampaniaRepository
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
 import javax.inject.Inject
 
 /**
@@ -10,7 +15,13 @@ import javax.inject.Inject
 class EliminarCampaniaUseCase @Inject constructor(
     private val campaniaRepository: CampaniaRepository
 ) {
-    suspend operator fun invoke(campania: Campania) {
-        campaniaRepository.deleteCampania(campania)
-    }
+    operator fun invoke(campania: Campania): Flow<Resource<Unit>> = flow {
+        emit(Resource.Loading)
+        try {
+            campaniaRepository.deleteCampania(campania)
+            emit(Resource.Success(Unit))
+        } catch (e: Exception) {
+            emit(Resource.Error(e.localizedMessage ?: "Error desconocido", e))
+        }
+    }.flowOn(Dispatchers.IO)
 }